website-25: move `careers` content to `join-the-team`
