+++ apps/website-25/src/components/join-the-team/CultureSection.tsx
