+++ apps/website-25/src/components/join-the-team/ValuesSection.tsx
